refactor(perf): split /coaching into server shell + motion island
Mirrors pattern from /music (Wave 17) and /golden-age (Wave 18).
Server component handles metadata + JSON-LD + static layout.
Client island holds Framer Motion + interactive state.
